# faster iandnot between bitmap and array containers