Faster and fewer allocs! 

```
name                                            old time/op    new time/op    delta
AndNot/inPlace=true/left=bitmap/right=array-10    12.9µs ±72%     7.3µs ±26%   -43.44%  (p=0.000 n=10+10)

name                                            old alloc/op   new alloc/op   delta
AndNot/inPlace=true/left=bitmap/right=array-10    8.22kB ± 0%    0.00kB       -100.00%  (p=0.000 n=10+10)

name                                            old allocs/op  new allocs/op  delta
AndNot/inPlace=true/left=bitmap/right=array-10      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)

```